move BlockHighlightContext into the page folder
